Refactor using existing methods
