+++ Jint/Runtime/Interop/InteropHelper.cs
@@ -298,4 +298,33 @@ internal sealed class InteropHelper
             yield return match;
         }
     }
+
+    internal static bool IsAwaitable(object? obj)
+    {
+        if (obj is null)
+        {
+            return false;
+        }
+        if (obj is Task)
+        {
+            return true;
+        }
+#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP
+            if (obj is ValueTask)
+            {
+                return true;
+            }
+
+            // ValueTask<T> is not derived from ValueTask, so we need to check for it explicitly
+            var type = obj.GetType();
+            if (!type.IsGenericType)
+            {
+                return false;
+            }
+
+            return type.GetGenericTypeDefinition() == typeof(ValueTask<>);
+#else
+        return false;
+#endif
+    }
 }
